+++ .changes/next-release/feature-AmazonS3-d8d7a87.json
@@ -1,0 +1,6 @@
+{
+    "type": "feature",
+    "category": "Amazon S3",
+    "contributor": "",
+    "description": "Add support for maxInFlightParts to multipart upload (PutObject) in MultipartS3AsyncClient."
+}
